// Reacts/Client/ToolGUI/ToolShape.ts
export default interface ToolShape
{
    id?:       string | null;
    x?:        number | null;
    y?:        number | null;
    circle?:   boolean | null;
    opacity?:  number | null;
    rgb?:      string | null;
    rotation?: number | null;
    scale?:    number | null;
    sides?:    number | null;
    star?:     boolean | null;
}